Fix `GetResourceLibraryProcAddress` to support LLVM linker.
This changes the way exported symbols are looked up to match the
way `GetProcAddress` does it, as per the explanation at
https://learn.microsoft.com/en-us/archive/msdn-magazine/2002/march/inside-windows-an-in-depth-look-into-the-win32-portable-executable-file-format-part-2.
The current method works for MSVC-linked plugins because the index of the
export name coincidentally always matches the index of the export address,
since MSVC always generates the export table that way. However, LLVM
will always make the export address index at 0 `NULL`, so the name and
address are off by one. This change allows the lookup to work for both
linkers.